+++ database/migrations/2020_05_22_080254_create_questions_table.php
@@ -18,14 +18,17 @@ class CreateQuestionsTable extends Migration
             $table->nullableMorphs('addedby'); // facilitator professional parent learner member
             $table->nullableMorphs('questionable'); // user conversation assessmentsection post discussion message
             $table->text('body');
-            $table->enum('state',['PENDING','ANSWERED','COMPLETE','SEEN','SENT','RECEIVED'])->nullable();
-            $table->mediumInteger('score_over')->nullable();
+            $table->enum('state', ['PENDING', 'ANSWERED', 'COMPLETE', 'SEEN', 'SENT', 'RECEIVED'])->nullable();
+            $table->float('score_over', 6, 2)->nullable();
             $table->text('hint')->nullable();
             $table->smallInteger('position')->nullable();
-            $table->enum('answer_type',[
-                'TRUE_FALSE','LONG_ANSWER','SHORT_ANSWER',
-                'IMAGE','VIDEO','AUDIO',
-                'OPTION','NUMBER','FLOW','ARRANGE']
+            $table->enum(
+                'answer_type',
+                [
+                    'TRUE_FALSE', 'LONG_ANSWER', 'SHORT_ANSWER',
+                    'IMAGE', 'VIDEO', 'AUDIO',
+                    'OPTION', 'NUMBER', 'FLOW', 'ARRANGE'
+                ]
             )->nullable();
             $table->json('correct_possible_answers')->nullable();
             $table->timestamp('published_at')->nullable();
